round-2 audit fixes: reset-offline guard, fill-cap corruption, bridge CSRF, cloud clamp
Two deep audits (fill state-machine + everything-else) found these; all fixed:

PLUGIN correctness:
- HIGH: resetJournal() run while logged out (null live offers) persisted an unseeded
  wipe → on next login in-flight offers re-booked their already-filled units, re-adding
  what was wiped. Now refuses with "log in first" when offers are unreadable.
- MED: the 3000-fill cap did a blind fills.remove(0) — could drop a BUY whose SELL was
  still logged, turning that sell into unmatched "pure profit" (overstated P&L) and
  vanishing the held lot. Replaced with pruneFills(): drop only FLAT (fully bought+sold)
  items, never orphaning; cap raised to 5000; atomic removeIf on the COW list.
- MED: upgrade-mode seed now unconditional on first sight (a pre-slotKey save loads
  key=null, so the old key-match guard let completed-uncollected offers double-book).
- LOW: delta guard gates on qty only (clamp spent≥0) so a lagging spend can't drop units;
  deleted the dead 5-arg saveFills overload; SetFlip.profit is long; escape item names in
  the HTML row tooltip.

SECURITY:
- MED: bridge POST now requires same-origin (Origin/Referer host == Host) — a drive-by
  site can no longer write your journal under the default empty token.

WEB / PWA:
- MED: sanitizeCfg() clamps every EXTERNAL config blob (cloud/bridge/localStorage) so a
  0/NaN cycleH or swingHoldHours can't spew Infinity gp/h in Deep/Slots.
- MED: web realistic-price guard now falls back to 1h VWAP (matches the plugin) so
  illiquid items price identically on phone and panel.
- generated real 192/512 PWA icons (Chrome install-prompt floor); sw.js caches shell files
  individually via allSettled (one missing asset no longer breaks the whole offline cache).

+++ runelite-plugin/src/main/java/com/geflip/GeflipScanner.java
@@ -239,7 +239,7 @@ class GeflipScanner
 	{
 		String name, dir;       // e.g. "Bandos armour set", "buy pieces → sell set"
 		long buyTotal, sellNet; // cost of the buy side, net-of-tax proceeds of the sell side
-		int profit;
+		long profit;
 	}
 
 	// Sets defined by NAME (resolved to ids via the live mapping, so no hardcoded/guessable ids;
@@ -301,13 +301,13 @@ class GeflipScanner
 			SetFlip f = new SetFlip();
 			f.name = s[0];
 			if (combine >= split && combine >= Math.max(1, cfg.minMargin()))
-			{ f.dir = "buy pieces → sell set"; f.buyTotal = piecesBuy; f.sellNet = setSell - saleTax(setSell, setMeta != null && setMeta.exempt); f.profit = (int) combine; }
+			{ f.dir = "buy pieces → sell set"; f.buyTotal = piecesBuy; f.sellNet = setSell - saleTax(setSell, setMeta != null && setMeta.exempt); f.profit = combine; }
 			else if (split >= Math.max(1, cfg.minMargin()))
-			{ f.dir = "buy set → sell pieces"; f.buyTotal = setBuy; f.sellNet = piecesSellNet; f.profit = (int) split; }
+			{ f.dir = "buy set → sell pieces"; f.buyTotal = setBuy; f.sellNet = piecesSellNet; f.profit = split; }
 			else continue;
 			out.add(f);
 		}
-		out.sort(Comparator.<SetFlip>comparingInt(x -> -x.profit));
+		out.sort(Comparator.<SetFlip>comparingLong(x -> -x.profit));
 		return out;
 	}
